% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DYNPROG.R
\name{optimal.cost}
\alias{optimal.cost}
\title{A helper function which determines the optimal cost for a given subsignal}
\usage{
optimal.cost(Q.vec, S.vec, start.index, end.index)
}
\arguments{
\item{Q.vec}{The cumulative sum of squares for the full signal}

\item{S.vec}{The cumulative sum for the full signal}

\item{start.index}{The first index in the subsignal, inclusive}

\item{end.index}{The last index in the subsignal, inclusive}
}
\value{
The optimal cost for the subsignal
}
\description{
A helper function which determines the optimal cost for a given subsignal
}
\examples{
signal <- c( runif(25, min=0, max=2),
             runif(25, min=2, max=5) )
Q.vec <- c( 0, cumsum( signal^2 ) )
S.vec <- c( 0, cumsum( signal ) )
cost <- optimal.cost( Q.vec, S.vec, 1, length( signal ) )
}
