% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Performs basic principal component analysis on the given data matrix}
\usage{
PCA(data.mat)
}
\arguments{
\item{data.mat}{The matrix of data to perform PCA on}
}
\value{
A list which contains the following elements is returned:
rotation: the matrix of weights for the original variables
lambda: the matrix of lambda values
}
\description{
Performs basic principal component analysis on the given data matrix
}
\examples{
test.matrix <- as.matrix( iris[ 1:50, 1:4 ] )
PCA.fit <- PCA( test.matrix )
PCA.fit$rotation
PCA.fit$lambda
}
